Added four new methods to Simulator
The first one is a getButton(int) this method takes an int and returns
the reference for the button of that index if it exists (throws
exception if not). Second method is getCell, same thing but with the
braille cells instead. I added these two because they allow the user to
actually create his own scenarios, he can now just call a specific
button and add an action listener or display a specific letter on any
specific cell. The third method is clearAllCells which just clears all
cells. The fourth one is displayString which goes through the cells in
order and displays the string, effectively spelling out a sentence in
Braille.