Wrote the constructor for the player app, take a look at it
You'll understand what I did if you read the code it's pretty
straightforward. The most important thing I added is the switch-case,
because that's what's going to be handling the different things the
player app should be able to do.
Every case corresponds to a thing the player app should handle, the
names of the methods should make it pretty obv which cases correspond to
what.
The methods take the scanner as imput because the tags (<a>, <t> etc.)
just indicate the BEGINING of the block of whatever.
So in out format every block will have to be terminated by another tag,
let's say <end>, and so since the methods take the scanner as input,
they will keep reading the next lines and doing what they do until the
next line is <end> at which point they just return and the main while
loop in the constructor keeps going.