I'm just going to copy-paste what I said on facebook
So there are two new tags (we can change the tags later) <cs> and <cf>
<cs> is choice success and <cf> is choice fail the idea is that after a
<c> tag there has to be both a <cs> and a <cf> tag, even if they're
empty, they just have to exist. In the case where the user enters the
correct button, <cs> is read and executed and <cf> is skipped. If he
enters the wrong answer <cs> is skipped and <cf> is read and executed.
Now the cool part is, you can just put another choice inside <cs> or
<cf>, so when the prof said "maybe if he enters a wrong answer you give
him an easier question instead", you can actually do that.

Added two additional tags after debugging.
In addition to <cs> and <cf> there's <break> and <end_c> now. The reason
is because I needed two tags in case at some point I wanted to skip
every single other branching path and go back to the main scenario. I
didn't notice it before because it's only a problem if you put choices
in choices, because when you enter a correct answer it would just
execute the next choice even if that one as conditionnal upon a
different branch. The reason is even though I had it coded in a way
where it skipped all consecutive <cs> and <cf>s, it didn't automatically
skip <c>s, since it considered them a new choice "tree" in the scenario.
With the new <break> tag, you add that whenever you want that specific
branch to be the "final" one, as in kick the program out of that tree
and coninute the scenario as normal. The <end_c> tag is just where the
<break> tag skips to. So, usage wise, you end every choice tree with ONE
<end_c> tag, only at the point where the normal scenario continues, and
whenever you want to skip TO that one, you use <break>.